+++ packages/firebase/src/model/userDevice/handlers/index.ts
@@ -1,0 +1,5 @@
+import addUserDevice from "./addUserDevice";
+
+export default {
+  addUserDevice,
+};